Organize folders; rename Objective-C to ObjC everywhere
