+++ src/T4TemplateWriter/CodeHelpers/Java/PropertyHelper.cs
