+++ src/T4TemplateWriter/CodeHelpers/Java/StringExtensions.cs
