+++ src/T4TemplateWriter/CodeHelpers/ObjC/PropertyHelper.cs
