+++ src/T4TemplateWriter/TemplateProcessor/CustomHost.cs
