+++ src/T4TemplateWriter/TemplateProcessor/ITemplateInfo.cs
@@ -1,10 +1,10 @@
 ﻿using System;
-namespace Vipr.T4TemplateWriter {
+namespace Vipr.T4TemplateWriter.TemplateProcessor {
     interface ITemplateInfo {
         string Id { get; }
         string TemplateLanguage { get; set; }
         string TemplateName { get; set; }
-        global::Vipr.T4TemplateWriter.TemplateType TemplateType { get; set; }
+        global::Vipr.T4TemplateWriter.TemplateProcessor.TemplateType TemplateType { get; set; }
 
     }
