+++ src/T4TemplateWriter/Templates/Java/Base/BaseTemplateCustomHost.java.tt
