+++ src/T4TemplateWriter/Templates/Java/Fetcher/EntityClient.java.tt
