+++ src/T4TemplateWriter/Templates/Java/Fetcher/EntityCollectionOperations.java.tt
