+++ src/T4TemplateWriter/Templates/Java/Fetcher/EntityFetcher.java.tt
