+++ src/T4TemplateWriter/Templates/Java/Fetcher/EntityOperations.java.tt
