+++ src/T4TemplateWriter/Templates/Java/Model/ComplexType.java.tt
