+++ src/T4TemplateWriter/Templates/Java/Model/EntityType.java.tt
