+++ src/T4TemplateWriter/Templates/Java/Model/EnumType.java.tt
