+++ src/T4TemplateWriter/Templates/ObjC/Base/BaseTemplate.m.tt
