+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityClient.h.tt
