+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityClient.m.tt
