+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityCollectionFetcher.h.tt
