+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityCollectionFetcher.m.tt
