+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityCollectionOperations.h.tt
