+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityCollectionOperations.m.tt
