+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityFetcher.h.tt
