+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityFetcher.m.tt
