+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityOperations.h.tt
