+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/EntityOperations.m.tt
