+++ src/T4TemplateWriter/Templates/ObjC/Fetcher/Fetchers.h.tt
