+++ src/T4TemplateWriter/Templates/ObjC/Model/EntityType.h.tt
