+++ src/T4TemplateWriter/Templates/ObjC/Model/EntityType.m.tt
