+++ src/T4TemplateWriter/Templates/ObjC/Model/Protocols.h.tt
